+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changed
 
+- [FORK] Add choice dialog for own story room (Issue #21)
 - [FORK] Rebranding to FF Chat (Issue #18)
 - [FORK] Set up and document Android build and testing process (Issue #11)
 - [FORK] Add solid ring indicator for unseen stories (Issue #27)